Added oColorTween, grouped FlashPunk objects, added oFadeOut, added scr_Extensions, began Dust event objects
Created an oColorTween for the oFadeOut object, along with the oFadeOut object itself.
Moved FlashPunk objects (only Tweens and a Matrix currently) to their own group inside the Objects group.
Added a scr_Extensions script to hold functions that can be used elsewhere but don't necessarily belong in a given object. Currently holds two helper functions for translating colors with RGB values as a single number, as well as an array reversal function needed for the leaving sequence.
Started work on the Dust/Sponge/Whiteboard objects but more research into Surfaces is needed to finish the Whiteboard object, which is needed for the other two objects in the event.
Added a Matrix object which may or may not be used for the Whiteboard. I will delete this if it goes unused.