Adjusted transition buffer frame
Changed the frame buffer for the leaving transition to use ratios of the scaled/original sizes for the camera view and application surface. The desired portion of the application surface is copied to a sprite, then the headstone image x/y scaling are set to the inverse ratio to de-scale it back down to the regular size.